requirements.txt
setuptools
pydeep2
